Netscaler: add awareness of netscalers running in High Availability clusters.
If talking to the non-active node in a cluster, changes will have no effect. This commit adds a 'resolve_primary' parameter which, if set, will make the module ask the netscaler for its HA configuration and use the primary node for subsequent connections (which may well be the same node).